Add LoginForm component: implement user authentication with magic link functionality and error handling, and refactor LoginPage to utilize the new component for improved code organization.
